Allow hermes-transform modifyInPlace API to modify the program node
Summary:
Previously the `context.modifyNodeInPlace` API worked via the `context.replaceNode` under the hood. However this meant the API did not work on the program node, since `context.replaceNode` does the replacement by walking to the node's parent then assigning the new node to that position, but program does not have a parent so this failed.

This meant that there were some basic things you just couldn't modify, for example the `docblock` comment can't be safely changed or the body of a program. Even worse if you happened to have an empty file there is no way you can add new statements.

This change creates a new dedicated "mutation" for `context.modifyNodeInPlace`. This mutation works by iterating through the new properties and just assigning them to the node. This way the parent node is not needed and so it will work on program.

To have this all work end to end i had to make some additional changes:
- Modified the comments API to allow us to correctly add any new docblock comments to the first statement.
- Changed the codegen to codegen helpers for the program node now they can be modified.

Reviewed By: SamChou19815

Differential Revision: D59309180

fbshipit-source-id: a5750928112fa95549ce4597b6435e3a36192c1a